% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{format}
\alias{format}
\title{Formatter}
\usage{
format(file, style = c("tidy", "grk"))
}
\arguments{
\item{file}{File to style.}

\item{style}{Style to apply.}
}
\description{
Formatter for conform.nvim
}
